+++ DKPlugins/DK/DKUtil.h
@@ -68,13 +68,13 @@ public:
 	static void SendTick();
 
 	//https://stackoverflow.com/questions/63166/how-to-determine-cpu-and-memory-consumption-from-inside-a-process
-	static bool VirtualMemory(float& virtualMemory);
-	static bool VirtualMemoryUsed(float& virtualMemory);
-	static bool VirtualMemoryUsedByApp(float& virtualMemory);
-	static bool PhysicalMemory(float& physicalMemory);
-	static bool PhysicalMemoryUsed(float& physicalMemory);
-	static bool PhysicalMemoryUsedByApp(float& physicalMemory);
-	static bool CpuUsed(int& cpu);
+	static bool VirtualMemory(unsigned long long& virtualMemory);
+	static bool VirtualMemoryUsed(unsigned long long& virtualMemory);
+	static bool VirtualMemoryUsedByApp(unsigned int& virtualMemory);
+	static bool PhysicalMemory(unsigned long long& physicalMemory);
+	static bool PhysicalMemoryUsed(unsigned long long& physicalMemory);
+	static bool PhysicalMemoryUsedByApp(unsigned int& physicalMemory);
+	static bool CpuUsed(long& cpu);
 	static bool CpuUsedByApp(int& cpu);
 
 	//Frame limiter
